streamlit_apps/workshop_ui/requirements.txt
streamlit>=1.28.0
google-adk>=1.17.0
google-genai>=1.17.0
python-dotenv>=1.0.1
